+++ examples/MNIST/image_file.sh
@@ -1,7 +1,7 @@
 cat > image.csv << EOF
 key, image
-four, image1.png
-five, image2.png
-three, image3.png
-two, image4.png
+four, image1.mnistpng
+five, image2.mnistpng
+three, image3.mnistpng
+two, image4.mnistpng
 EOF