+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Docstrings and doctestable examples to `record.py`.
 - Inputs can be validated using operations
   - `validate` parameter in `Input` takes `Operation.instance_name`
+- Test tensorflow DNNEstimator documentation exaples in CI
+- Add python code for tensorflow DNNEstimator
 ### Fixed
 - New model tutorial mentions file paths that should be edited.