+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Docstrings and doctestable examples to `record.py`.
 - Inputs can be validated using operations
-- `validate` parameter in `Input` takes `Operation.instance_name`
+  - `validate` parameter in `Input` takes `Operation.instance_name`
 - Logistic Regression with SAG optimizer
 
 ### Fixed