+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Docstrings and doctestable examples to `record.py`.
+- Inputs can be validated using operations
+  - `validate` parameter in `Input` takes `Operation.instance_name`
+- Logistic Regression with SAG optimizer
+- Test tensorflow DNNEstimator documentation exaples in CI
+- Add python code for tensorflow DNNEstimator
 ### Fixed
 - New model tutorial mentions file paths that should be edited.